Major Fix: Updated Plug-In to Meet Discussed Functionality
Added plug-in.tar.gz. This contains the plug-in functionalities for the server. Sent to the user's home directory on the server when the CMR is run without any models found. It allows users to easily install models on the server-end.

Updated the client end functionality. Client will be given the available models on a given machine and allowed to access them via the CMR.

Removed functionality to send source code through agave. Users must now use the plug-in sent to their home directory to install models. They must also be within the singularity image.